Refactoring code and set template
