+++ Assets/Scripts/Characters/PlayerController.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Animations;
 
 /// <summary>
 /// Base Class that all characters in this game must inherit from
