+++ Assets/Scripts/Enemies/EnemyGrunt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f19762e4192fb87418904ad4cf119c11
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
